requirements.txt
requests==2.32.3
regex==2024.11.6
logging==0.4.9.6
datetime==5.5
python-dotenv==1.0.1
shapely==2.0.6
marshmallow==3.24.1
flask-marshmallow==1.3.0
azure-functions==1.21.3